agent: Break up agent utils
