+++ agent/src/main/java/com/code_intelligence/jazzer/agent/BUILD.bazel
@@ -11,6 +11,8 @@ kt_jvm_library(
     deps = [
         "//agent/src/main/java/com/code_intelligence/jazzer/instrumentor",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime",
+        "//agent/src/main/java/com/code_intelligence/jazzer/utils:class_name_globber",
+        "//agent/src/main/java/com/code_intelligence/jazzer/utils:manifest_utils",
         "//driver/src/main/java/com/code_intelligence/jazzer/driver:opt",
     ],
 )
