+++ agent/src/main/java/com/code_intelligence/jazzer/instrumentor/BUILD.bazel
@@ -26,6 +26,7 @@ kt_jvm_library(
     deps = [
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime",
         "//agent/src/main/java/com/code_intelligence/jazzer/utils",
+        "//agent/src/main/java/com/code_intelligence/jazzer/utils:class_name_globber",
         "@com_github_classgraph_classgraph//:classgraph",
         "@com_github_jetbrains_kotlin//:kotlin-reflect",
         "@jazzer_jacoco//:jacoco_internal",
