+++ agent/src/main/java/com/code_intelligence/jazzer/runtime/BUILD.bazel
@@ -21,9 +21,7 @@ java_jni_library(
     visibility = [
         "//driver/src/main/native/com/code_intelligence/jazzer/driver:__pkg__",
     ],
-    deps = [
-        "//agent/src/main/java/com/code_intelligence/jazzer/utils",
-    ],
+    deps = ["//agent/src/main/java/com/code_intelligence/jazzer/utils"],
 )
 
 java_library(
@@ -50,6 +48,5 @@ java_library(
         ":coverage_map",
         ":trace_data_flow_native_callbacks",
         "//agent/src/main/java/com/code_intelligence/jazzer/api",
-        "//agent/src/main/java/com/code_intelligence/jazzer/utils",
     ],
 )
