+++ driver/src/main/java/com/code_intelligence/jazzer/autofuzz/BUILD.bazel
@@ -11,6 +11,7 @@ java_library(
     deps = [
         "//agent/src/main/java/com/code_intelligence/jazzer/api",
         "//agent/src/main/java/com/code_intelligence/jazzer/utils",
+        "//agent/src/main/java/com/code_intelligence/jazzer/utils:simple_glob_matcher",
         "@com_github_classgraph_classgraph//:classgraph",
         "@com_github_jhalterman_typetools//:typetools",
     ],
