+++ driver/src/main/java/com/code_intelligence/jazzer/driver/BUILD.bazel
@@ -1,4 +1,5 @@
 load("@fmeum_rules_jni//jni:defs.bzl", "java_jni_library")
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
 
 java_library(
     name = "driver",
@@ -15,6 +16,12 @@ java_library(
     ],
 )
 
+kt_jvm_library(
+    name = "exception_utils",
+    srcs = ["ExceptionUtils.kt"],
+    deps = ["//agent/src/main/java/com/code_intelligence/jazzer/api"],
+)
+
 java_jni_library(
     name = "fuzz_target_runner",
     srcs = ["FuzzTargetRunner.java"],
@@ -27,6 +34,7 @@ java_jni_library(
         "//driver/src/test:__subpackages__",
     ],
     deps = [
+        ":exception_utils",
         ":fuzzed_data_provider_impl",
         ":opt",
         ":recording_fuzzed_data_provider",
@@ -38,7 +46,7 @@ java_jni_library(
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime:coverage_map",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime:unsafe_provider",
-        "//agent/src/main/java/com/code_intelligence/jazzer/utils",
+        "//agent/src/main/java/com/code_intelligence/jazzer/utils:manifest_utils",
         "//driver/src/main/java/com/code_intelligence/jazzer/autofuzz",
     ],
 )
